requirements.txt
voxelbotutils
discord-py-slash-command